Fix smap <Plug>(vsnip-expand)
A recent commit introduced a regression for the newly introduced
<Plug>(vsnip-expand) mapping.
I've changed a few spots to make this mapping more robust.

The context for the snippet was detected incorrectly as the current
column was at the beginning of the selection and no snippet could be
detected.

Now the selected text will have the cursor at the end instead of the
front.
This change will probably affect no user and eases the
implementation.

Also the column offset in select mode is 1 instead of 2.

Now the implementation of <Plug>(vsnip-expand) should be fully supported
instead of "just works".