Nvim 0.6: fix snippet detection in start packages (#224)
Nvim 0.6 no longer adds start packages to &rtp.

This means that, for example, simply installing friendly-snippets as a
start package no longer works. It'll instead need to be installed as an
opt package and packadded so it appears in &rtp for vim-vsnip to detect
it (some package managers may do this indirectly).

The recommended approach is to use nvim_list_runtime_paths instead; use
it if it's available.